requirements_production.txt
# Dependências básicas do sistema
Flask
openpyxl
Werkzeug==2.0.1
Jinja2==3.0.1
itsdangerous==2.0.1
click==8.0.1
fpdf2
matplotlib

# Dependências adicionais para produção
waitress

# Para criar distribuição executável
pyinstaller

# Para interface desktop nativa
pywebview
